AD-51 Moved Ballot model and services to their own namespace
