+++ TrueVote.Api.Tests/ServiceTests/BallotTest.cs
@@ -1,0 +1,70 @@
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
+using MockQueryable.Moq;
+using Moq;
+using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Net;
+using System.Text;
+using System.Threading.Tasks;
+using TrueVote.Api.Models;
+using TrueVote.Api.Services;
+using TrueVote.Api.Tests.Helpers;
+using Xunit;
+using Xunit.Abstractions;
+
+namespace TrueVote.Api.Tests.ServiceTests
+{
+    public class BallotTest : TestHelper
+    {
+        public BallotTest(ITestOutputHelper output) : base(output)
+        {
+        }
+
+        [Fact]
+        public async Task SubmitsBallot()
+        {
+            var baseBallotObj = new SubmitBallotModel { ElectionId = "68" };
+            var byteArray = Encoding.ASCII.GetBytes(JsonConvert.SerializeObject(baseBallotObj));
+            _httpContext.Request.Body = new MemoryStream(byteArray);
+
+            var ret = await _ballotApi.SubmitBallot(_httpContext.Request) as CreatedResult;
+            Assert.NotNull(ret);
+            var objectResult = Assert.IsType<CreatedResult>(ret);
+            Assert.Equal((int) HttpStatusCode.Created, objectResult.StatusCode);
+
+            var val = ret.Value as SubmitBallotModelResponse;
+            Assert.NotNull(val);
+
+            _output.WriteLine($"Item: {val}");
+
+            _output.WriteLine($"Item.Message: {val.Message}");
+            _output.WriteLine($"Item.ElectionId: {val.ElectionId}");
+
+            Assert.Contains("Ballot successfully submitted.", val.Message);
+            Assert.NotEmpty(val.ElectionId);
+
+            _logHelper.Verify(LogLevel.Information, Times.Exactly(1));
+            _logHelper.Verify(LogLevel.Debug, Times.Exactly(2));
+        }
+
+        [Fact]
+        public async Task HandlesSubmitBallotError()
+        {
+            var baseBallotObj = "blah";
+            var byteArray = Encoding.ASCII.GetBytes(baseBallotObj);
+            _httpContext.Request.Body = new MemoryStream(byteArray);
+
+            var ret = await _ballotApi.SubmitBallot(_httpContext.Request);
+            Assert.NotNull(ret);
+            var objectResult = Assert.IsType<BadRequestObjectResult>(ret);
+            Assert.Equal((int) HttpStatusCode.BadRequest, objectResult.StatusCode);
+
+            _logHelper.Verify(LogLevel.Error, Times.Exactly(1));
+            _logHelper.Verify(LogLevel.Debug, Times.Exactly(2));
+        }
+    }
+}
