+++ TrueVote.Api.Tests/ServiceTests/ElectionTest.cs
@@ -301,48 +301,5 @@ namespace TrueVote.Api.Tests.ServiceTests
 
             _logHelper.Verify(LogLevel.Debug, Times.Exactly(1));
         }
-
-        [Fact]
-        public async Task SubmitsBallot()
-        {
-            var baseBallotObj = new SubmitBallotModel { ElectionId = "68" };
-            var byteArray = Encoding.ASCII.GetBytes(JsonConvert.SerializeObject(baseBallotObj));
-            _httpContext.Request.Body = new MemoryStream(byteArray);
-
-            var ret = await _electionApi.SubmitBallot(_httpContext.Request) as CreatedResult;
-            Assert.NotNull(ret);
-            var objectResult = Assert.IsType<CreatedResult>(ret);
-            Assert.Equal((int) HttpStatusCode.Created, objectResult.StatusCode);
-
-            var val = ret.Value as SubmitBallotModelResponse;
-            Assert.NotNull(val);
-
-            _output.WriteLine($"Item: {val}");
-
-            _output.WriteLine($"Item.Message: {val.Message}");
-            _output.WriteLine($"Item.ElectionId: {val.ElectionId}");
-
-            Assert.Contains("Ballot successfully submitted.", val.Message);
-            Assert.NotEmpty(val.ElectionId);
-
-            _logHelper.Verify(LogLevel.Information, Times.Exactly(1));
-            _logHelper.Verify(LogLevel.Debug, Times.Exactly(2));
-        }
-
-        [Fact]
-        public async Task HandlesSubmitBallotError()
-        {
-            var baseBallotObj = "blah";
-            var byteArray = Encoding.ASCII.GetBytes(baseBallotObj);
-            _httpContext.Request.Body = new MemoryStream(byteArray);
-
-            var ret = await _electionApi.SubmitBallot(_httpContext.Request);
-            Assert.NotNull(ret);
-            var objectResult = Assert.IsType<BadRequestObjectResult>(ret);
-            Assert.Equal((int) HttpStatusCode.BadRequest, objectResult.StatusCode);
-
-            _logHelper.Verify(LogLevel.Error, Times.Exactly(1));
-            _logHelper.Verify(LogLevel.Debug, Times.Exactly(2));
-        }
     }
 }
