+++ TrueVote.Api/Models/BallotModel.cs
@@ -1,0 +1,45 @@
+using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Attributes;
+using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Enums;
+using Newtonsoft.Json;
+using System.ComponentModel.DataAnnotations;
+using System.Diagnostics.CodeAnalysis;
+
+namespace TrueVote.Api.Models
+{
+    [ExcludeFromCodeCoverage]
+    public class SubmitBallotModel {
+        [OpenApiSchemaVisibility(OpenApiVisibilityType.Important)]
+        [OpenApiProperty(Description = "Election Id")]
+        [MaxLength(2048)]
+        [DataType(DataType.Text)]
+        [RegularExpression(Constants.GenericStringRegex)]
+        [JsonProperty(PropertyName = "ElectionId")]
+        [Key]
+        public string ElectionId { get; set; }
+
+        //[OpenApiSchemaVisibility(OpenApiVisibilityType.Important)]
+        //[OpenApiProperty(Description = "Election Model")]
+        //[DataType("ElectionModel")]
+        //[JsonProperty(PropertyName = "ElectionModel", Required = Required.Always)]
+        //public ElectionModel ElectionModel { get; set; }
+    }
+
+    [ExcludeFromCodeCoverage]
+    public class SubmitBallotModelResponse {
+        [OpenApiSchemaVisibility(OpenApiVisibilityType.Important)]
+        [OpenApiProperty(Description = "Election Id")]
+        [MaxLength(2048)]
+        [DataType(DataType.Text)]
+        [RegularExpression(Constants.GenericStringRegex)]
+        [JsonProperty(PropertyName = "ElectionId")]
+        public string ElectionId { get; set; }
+
+        [OpenApiSchemaVisibility(OpenApiVisibilityType.Important)]
+        [OpenApiProperty(Description = "Message")]
+        [MaxLength(32768)]
+        [DataType(DataType.Text)]
+        [RegularExpression(Constants.GenericStringRegex)]
+        [JsonProperty(PropertyName = "Message")]
+        public string Message { get; set; }
+    }
+}
