+++ TrueVote.Api/Models/ElectionModel.cs
@@ -236,42 +236,4 @@ namespace TrueVote.Api.Models
         [JsonProperty(PropertyName = "RaceIds", Required = Required.Always)]
         public List<string> RaceIds { get; set; }
     }
-
-
-    [ExcludeFromCodeCoverage]
-    public class SubmitBallotModel {
-        [OpenApiSchemaVisibility(OpenApiVisibilityType.Important)]
-        [OpenApiProperty(Description = "Election Id")]
-        [MaxLength(2048)]
-        [DataType(DataType.Text)]
-        [RegularExpression(Constants.GenericStringRegex)]
-        [JsonProperty(PropertyName = "ElectionId")]
-        [Key]
-        public string ElectionId { get; set; }
-
-        //[OpenApiSchemaVisibility(OpenApiVisibilityType.Important)]
-        //[OpenApiProperty(Description = "Election Model")]
-        //[DataType("ElectionModel")]
-        //[JsonProperty(PropertyName = "ElectionModel", Required = Required.Always)]
-        //public ElectionModel ElectionModel { get; set; }
-    }
-
-    [ExcludeFromCodeCoverage]
-    public class SubmitBallotModelResponse {
-        [OpenApiSchemaVisibility(OpenApiVisibilityType.Important)]
-        [OpenApiProperty(Description = "Election Id")]
-        [MaxLength(2048)]
-        [DataType(DataType.Text)]
-        [RegularExpression(Constants.GenericStringRegex)]
-        [JsonProperty(PropertyName = "ElectionId")]
-        public string ElectionId { get; set; }
-
-        [OpenApiSchemaVisibility(OpenApiVisibilityType.Important)]
-        [OpenApiProperty(Description = "Message")]
-        [MaxLength(32768)]
-        [DataType(DataType.Text)]
-        [RegularExpression(Constants.GenericStringRegex)]
-        [JsonProperty(PropertyName = "Message")]
-        public string Message { get; set; }
-    }
 }
